requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
black==25.1.0
cachetools==5.5.2
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
distlib==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
flake8==7.3.0
google==3.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jiter==0.10.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
openai==1.99.5
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
rsa==4.9.1
ruff==0.12.8
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.33.1
pytest-asyncio
